+++ src/components/popover/popover.module.scss
@@ -1,9 +1,31 @@
 @import 'colors';
+@import 'variables';
 
 .root {
-  width: 100%;
+  width: 100vw;
   height: 100vh;
   position: absolute;
   z-index: 99;
+  top: 0;
+  left: 0;
+  overflow: hidden;
+  background-color: rgba($color: $color-dark, $alpha: 0.1);
+}
+
+.menu {
+  position: absolute;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 40%;
   bottom: 0;
+  background-color: $color-white;
+  border-top-left-radius: $border-radius;
+  border-top-right-radius: $border-radius;
+  transition: transform 200ms;
+  &.closed {
+    transform: translateY(100%);
+  }
 }
