--- src/components/selector/selector.module.scss
@@ -1,36 +1,0 @@
-@import 'colors';
-@import 'variables';
-
-.selector {
-  border: solid 1px $color-neutral-dark;
-  padding: 0.5rem;
-  text-align: center;
-  background-color: $color-overlay-disabled;
-  width: 5rem;
-  margin-bottom: 1rem;
-  @include md() {
-    width: 6rem;
-  }
-  &.active {
-    background-color: $color-overlay;
-  }
-}
-
-.sprite {
-  height: 2rem;
-  @include md() {
-    height: 3rem;
-  }
-}
-
-.healthbar {
-  border: solid 1px $color-neutral-dark;
-  height: 1rem;
-  width: 100%;
-  background-color: $color-neutral-light;
-}
-
-.health {
-  height: 100%;
-}
-
