+++ src/components/switch/Switch.tsx
@@ -1,8 +1,9 @@
 import { TeamMember } from "@adibkhan/pogo-web-backend"
 import style from './switch.module.scss'
 import { useEffect, useState } from "react"
-import Selector from '../selector/Selector'
 import classnames from "classnames"
+import getColor from "@common/actions/getColor"
+import { SERVER } from "@config/index"
 
 interface SwitchProps {
   team: TeamMember[],
@@ -15,6 +16,47 @@ interface SwitchProps {
 interface CharArray {
   char: TeamMember,
   index: number
+}
+
+interface SelectorProps {
+  member: TeamMember,
+  index: number,
+  active: boolean,
+  onClick: (to: number) => void,
+}
+
+const getImage = (sid: number): string => {
+  return `${SERVER}/mini/${sid}.png`
+}
+
+const Selector: React.FC<SelectorProps> = ({ member, index, active, onClick }) => {
+  const handleClick = (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
+    event.preventDefault()
+    event.stopPropagation()
+    onClick(index)
+  }
+
+  const image = getImage(member.sid)
+  const ratio = member.current ? Math.min(member.current.hp/member.hp, 1) : 1
+  const color = getColor(ratio)
+
+  return (
+    <button
+      onClick={handleClick}
+      className={classnames([style.selector, {
+        [style.active]: active
+      }])}>
+      <small>CP {member.cp}</small>
+      <br />
+      <img className={style.sprite} src={image} alt={member.speciesName}/>
+      <div
+        className={style.healthbar}
+        style={{ width: `${ratio*100}%`, backgroundColor: color }}
+      >
+        <div className={style.health} />
+      </div>
+    </button>
+  )
 }
 
 const Switch: React.FC<SwitchProps> = ({ team, pointer, countdown, onClick, modal }) => {
@@ -56,7 +98,10 @@ const Switch: React.FC<SwitchProps> = ({ team, pointer, countdown, onClick, moda
           )
         })
       }
-      <strong>{countdown}</strong>
+      {
+        !modal &&
+        <strong>{countdown}</strong>
+      }
     </div>
   )
 }
