+++ src/components/switch/switch.module.scss
@@ -1,3 +1,39 @@
+@import 'colors';
+@import 'variables';
+
+.selector {
+  border: solid 1px $color-neutral-dark;
+  padding: 0.5rem;
+  text-align: center;
+  background-color: $color-overlay-disabled;
+  width: 5rem;
+  margin-bottom: 1rem;
+  @include md() {
+    width: 6rem;
+  }
+  &.active {
+    background-color: $color-overlay;
+  }
+}
+
+.sprite {
+  height: 2rem;
+  @include md() {
+    height: 3rem;
+  }
+}
+
+.healthbar {
+  border: solid 1px $color-neutral-dark;
+  height: 1rem;
+  width: 100%;
+  background-color: $color-neutral-light;
+}
+
+.health {
+  height: 100%;
+}
+
 .root {
   background-color: transparent;
   position: absolute;
@@ -14,5 +50,15 @@
     flex-direction: row;
     justify-content: space-around;
     align-items: center;
+    position: relative;
+    width: 100%;
+
+    .selector {
+      margin-right: 1em;
+      margin-bottom: 0;
+      .sprite {
+        height: 3rem;
+      }
+    }
   }
 }
